Move all repositories to same package and manually define their beans
