+++ src/main/java/uk/gov/companieshouse/api/testdata/config/MongoConfig.java
@@ -1,14 +1,24 @@
 package uk.gov.companieshouse.api.testdata.config;
 
+import java.io.Serializable;
+
 import org.springframework.boot.autoconfigure.mongo.MongoProperties;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.SimpleMongoDbFactory;
+import org.springframework.data.mongodb.repository.support.MongoRepositoryFactoryBean;
+import org.springframework.data.repository.Repository;
 
 import com.mongodb.MongoClient;
 import com.mongodb.MongoClientURI;
+
+import uk.gov.companieshouse.api.testdata.repository.CompanyAuthCodeRepository;
+import uk.gov.companieshouse.api.testdata.repository.CompanyProfileRepository;
+import uk.gov.companieshouse.api.testdata.repository.FilingHistoryRepository;
+import uk.gov.companieshouse.api.testdata.repository.OfficerRepository;
+import uk.gov.companieshouse.api.testdata.repository.PersonsWithSignificantControlRepository;
 
 @Configuration
 @EnableConfigurationProperties(MongoProperties.class)
@@ -21,33 +31,42 @@ public class MongoConfig {
         this.mongoProperties = mongoProperties;
     }
 
-    @Bean(name = "companyProfileMongoTemplate")
-    public MongoTemplate companyProfileMongoTemplate() {
-        return createMongoTemplate(this.mongoProperties.getUri(), "company_profile");
+    @Bean
+    public CompanyProfileRepository companyProfileRepository() {
+        return getMongoRepositoryBean(CompanyProfileRepository.class, "company_profile");
     }
 
-    @Bean(name = "accountMongoTemplate")
-    public MongoTemplate accountMongoTemplate() {
-        return createMongoTemplate(this.mongoProperties.getUri(), "account");
+    @Bean
+    public CompanyAuthCodeRepository accountRepository() {
+        return getMongoRepositoryBean(CompanyAuthCodeRepository.class, "account");
     }
 
-    @Bean(name = "filingHistoryMongoTemplate")
-    public MongoTemplate filingHistoryMongoTemplate() {
-        return createMongoTemplate(this.mongoProperties.getUri(), "company_filing_history");
+    @Bean
+    public FilingHistoryRepository filingHistoryRepository() {
+        return getMongoRepositoryBean(FilingHistoryRepository.class, "company_filing_history");
     }
 
-    @Bean(name = "officerMongoTemplate")
-    public MongoTemplate officerMongoTemplate() {
-        return createMongoTemplate(this.mongoProperties.getUri(), "officer_appointments");
+    @Bean
+    public OfficerRepository officerRepository() {
+        return getMongoRepositoryBean(OfficerRepository.class, "officer_appointments");
     }
 
-    @Bean(name = "pscMongoTemplate")
-    public MongoTemplate pscMongoTemplate() {
-        return createMongoTemplate(this.mongoProperties.getUri(), "company_pscs");
+    @Bean
+    public PersonsWithSignificantControlRepository pscRepository() {
+        return getMongoRepositoryBean(PersonsWithSignificantControlRepository.class, "company_pscs");
     }
 
-    private MongoTemplate createMongoTemplate(final String uri, final String database) {
-        return new MongoTemplate(new SimpleMongoDbFactory(new MongoClient(new MongoClientURI(uri)), database));
+    private MongoTemplate createMongoTemplate(final String database) {
+        return new MongoTemplate(
+                new SimpleMongoDbFactory(new MongoClient(new MongoClientURI(this.mongoProperties.getUri())), database));
+    }
+
+    private <T extends Repository<S, ID>, S, ID extends Serializable> T getMongoRepositoryBean(Class<T> repositoryClass,
+            String database) {
+        MongoRepositoryFactoryBean<T, S, ID> mongoDbFactoryBean = new MongoRepositoryFactoryBean<>(repositoryClass);
+        mongoDbFactoryBean.setMongoOperations(createMongoTemplate(database));
+        mongoDbFactoryBean.afterPropertiesSet();
+        return mongoDbFactoryBean.getObject();
     }
 
 }
