--- src/main/java/uk/gov/companieshouse/api/testdata/config/PscMongoConfig.java
@@ -1,10 +1,0 @@
-package uk.gov.companieshouse.api.testdata.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
-
-@Configuration
-@EnableMongoRepositories(basePackages = "uk.gov.companieshouse.api.testdata.repository.psc", mongoTemplateRef = "pscMongoTemplate")
-public class PscMongoConfig {
-
-}
