+++ src/main/java/uk/gov/companieshouse/api/testdata/repository/FilingHistoryRepository.java
@@ -1,9 +1,11 @@
-package uk.gov.companieshouse.api.testdata.repository.filinghistory;
+package uk.gov.companieshouse.api.testdata.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.data.repository.NoRepositoryBean;
 
 import uk.gov.companieshouse.api.testdata.model.entity.FilingHistory;
 
+@NoRepositoryBean
 public interface FilingHistoryRepository extends MongoRepository<FilingHistory, String> {
 
     FilingHistory findByCompanyNumber(String companyId);
