+++ src/main/java/uk/gov/companieshouse/api/testdata/service/impl/OfficerListServiceImpl.java
@@ -18,7 +18,7 @@ import uk.gov.companieshouse.api.testdata.model.entity.DateOfBirth;
 import uk.gov.companieshouse.api.testdata.model.entity.Links;
 import uk.gov.companieshouse.api.testdata.model.entity.Officer;
 import uk.gov.companieshouse.api.testdata.model.entity.OfficerItem;
-import uk.gov.companieshouse.api.testdata.repository.officer.OfficerRepository;
+import uk.gov.companieshouse.api.testdata.repository.OfficerRepository;
 import uk.gov.companieshouse.api.testdata.service.DataService;
 import uk.gov.companieshouse.api.testdata.service.TestDataHelperService;
