+++ src/test/java/uk/gov/companieshouse/api/testdata/service/CompanyAuthCodeServiceImplTest.java
@@ -10,7 +10,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.companieshouse.api.testdata.exception.DataException;
 import uk.gov.companieshouse.api.testdata.exception.NoDataFoundException;
 import uk.gov.companieshouse.api.testdata.model.entity.CompanyAuthCode;
-import uk.gov.companieshouse.api.testdata.repository.account.CompanyAuthCodeRepository;
+import uk.gov.companieshouse.api.testdata.repository.CompanyAuthCodeRepository;
 import uk.gov.companieshouse.api.testdata.service.impl.CompanyAuthCodeServiceImpl;
 
 import java.util.Optional;
