+++ src/test/java/uk/gov/companieshouse/api/testdata/service/CompanyProfileServiceImplTest.java
@@ -10,7 +10,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.companieshouse.api.testdata.exception.DataException;
 import uk.gov.companieshouse.api.testdata.exception.NoDataFoundException;
 import uk.gov.companieshouse.api.testdata.model.entity.CompanyProfile;
-import uk.gov.companieshouse.api.testdata.repository.companyprofile.CompanyProfileRepository;
+import uk.gov.companieshouse.api.testdata.repository.CompanyProfileRepository;
 import uk.gov.companieshouse.api.testdata.service.impl.CompanyProfileServiceImpl;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
