+++ src/test/java/uk/gov/companieshouse/api/testdata/service/FilingHistoryServiceImplTest.java
@@ -10,7 +10,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.companieshouse.api.testdata.exception.DataException;
 import uk.gov.companieshouse.api.testdata.exception.NoDataFoundException;
 import uk.gov.companieshouse.api.testdata.model.entity.FilingHistory;
-import uk.gov.companieshouse.api.testdata.repository.filinghistory.FilingHistoryRepository;
+import uk.gov.companieshouse.api.testdata.repository.FilingHistoryRepository;
 import uk.gov.companieshouse.api.testdata.service.impl.FilingHistoryServiceImpl;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
