+++ src/test/java/uk/gov/companieshouse/api/testdata/service/OfficerListServiceImplTest.java
@@ -10,7 +10,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.companieshouse.api.testdata.exception.DataException;
 import uk.gov.companieshouse.api.testdata.exception.NoDataFoundException;
 import uk.gov.companieshouse.api.testdata.model.entity.Officer;
-import uk.gov.companieshouse.api.testdata.repository.officer.OfficerRepository;
+import uk.gov.companieshouse.api.testdata.repository.OfficerRepository;
 import uk.gov.companieshouse.api.testdata.service.impl.OfficerListServiceImpl;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
